issue 160 admin menu not allow delete
git-svn-id: http://xe-core.googlecode.com/svn/branches/1.5.0@9160 201d5d3c-b55e-5fd7-737f-ddc643e51545

+++ modules/menu/menu.admin.controller.php
@@ -66,6 +66,13 @@
          **/
         function procMenuAdminDelete() {
             $menu_srl = Context::get('menu_srl');
+
+			$oMenuAdminModel = &getAdminModel('menu');
+			$menu_info = $oMenuAdminModel->getMenu($menu_srl);
+
+			if($menu_info->title == '__XE_ADMIN__')
+				return new Object(-1, 'msg_adminmenu_cannot_delete');
+
             $this->deleteMenu($menu_srl);
 
 			$this->setMessage('success_deleted', 'info');
